backend/requirements.txt
amqp==5.3.1
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cssselect==1.2.0
Django==5.1.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-environ==0.11.2
django-timezone-field==7.0
djangorestframework==3.15.2
fake-useragent==1.5.1
feedparser==6.0.11
finnhub-python==2.4.21
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
kombu==5.4.2
lxml==5.3.0
mysqlclient==2.2.5
numpy==2.1.3
pandas==2.2.3
parse==1.20.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyee==11.1.1
PyMySQL==1.1.1
pyppeteer==2.0.0
pyquery==2.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
requests==2.32.3
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4

zipp==3.21.0
